youtube-video-qa-bot/requirements.txt
youtube-transcript-api
langchain-core
langchain-community
langchain-openai
langchain-text-splitters
openai
chromadb
python-dotenv
tiktoken